chore(suite-native): code reorg
