+++ suite-native/coin-enabling/src/coinEnablingSchema.ts
@@ -1,13 +1,6 @@
 import { yup } from '@suite-common/validators';
-import { type NetworkSymbol, networkSymbolCollection } from '@suite-common/wallet-config';
 
-export type EnabledCoins = Partial<Record<NetworkSymbol, boolean>>;
-
-export type EnabledCoinFieldName = `enabledCoins.${NetworkSymbol}`;
-
-export type CoinEnablingFormValues = {
-    enabledCoins: EnabledCoins;
-};
+import { type EnabledCoins } from './coinEnablingFormUtils';
 
 export const coinEnablingFormValidationSchema = yup.object({
     enabledCoins: yup
@@ -16,18 +9,3 @@ export const coinEnablingFormValidationSchema = yup.object({
             Object.values(value ?? {}).some(Boolean),
         ),
 });
-
-export const getEnabledCoinsFromNetworkSymbols = (symbols: NetworkSymbol[]): EnabledCoins =>
-    symbols.reduce<EnabledCoins>(
-        (enabledCoins, symbol) => ({
-            ...enabledCoins,
-            [symbol]: true,
-        }),
-        {},
-    );
-
-export const getEnabledCoinFieldName = (symbol: NetworkSymbol): EnabledCoinFieldName =>
-    `enabledCoins.${symbol}`;
-
-export const getNetworkSymbolsFromEnabledCoins = (enabledCoins: EnabledCoins): NetworkSymbol[] =>
-    networkSymbolCollection.filter(symbol => enabledCoins[symbol]);
